# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Moneyball Scout API",
  "description": "Live football data from API-Football, match ratings, form, valuations and manager ratings.",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/fixtures": {
   "get": {
    "tags": [
     "fixtures"
    ],
    "summary": "List Fixtures",
    "operationId": "list_fixtures_api_fixtures_get",
    "parameters": [
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "upcoming",
        "played"
       ],
       "type": "string",
       "default": "played",
       "title": "Status"
      }
     },
     {
      "name": "team_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Team Id"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 400,
       "default": 20,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/FixtureRow"
         },
         "title": "Response List Fixtures Api Fixtures Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/fixtures/{fixture_id}": {
   "get": {
    "tags": [
     "fixtures"
    ],
    "summary": "Fixture Detail",
    "operationId": "fixture_detail_api_fixtures__fixture_id__get",
    "parameters": [
     {
      "name": "fixture_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Fixture Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/FixtureDetail"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "FixtureDetail": {
    "properties": {
     "fixture": {
      "$ref": "#/components/schemas/FixtureRow"
     },
     "lineups": {
      "additionalProperties": {
       "items": {
        "$ref": "#/components/schemas/LineupPlayer"
       },
       "type": "array"
      },
      "type": "object",
      "title": "Lineups"
     }
    },
    "type": "object",
    "required": [
     "fixture",
     "lineups"
    ],
    "title": "FixtureDetail"
   },
   "FixtureRow": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "date": {
      "type": "string",
      "format": "date-time",
      "title": "Date"
     },
     "round": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Round"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "home": {
      "$ref": "#/components/schemas/TeamRef"
     },
     "away": {
      "$ref": "#/components/schemas/TeamRef"
     },
     "home_goals": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Home Goals"
     },
     "away_goals": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Away Goals"
     },
     "prediction": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Prediction"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "type": "object",
    "required": [
     "id",
     "date",
     "status",
     "home",
     "away"
    ],
    "title": "FixtureRow"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "LineupPlayer": {
    "properties": {
     "player_id": {
      "type": "integer",
      "title": "Player Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "photo": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Photo"
     },
     "position": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Position"
     },
     "minutes": {
      "type": "integer",
      "title": "Minutes"
     },
     "substitute": {
      "type": "boolean",
      "title": "Substitute"
     },
     "rating": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rating"
     },
     "api_rating": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Api Rating"
     },
     "components": {
      "additionalProperties": {
       "type": "number"
      },
      "type": "object",
      "title": "Components",
      "default": {}
     },
     "goals": {
      "type": "integer",
      "title": "Goals"
     },
     "assists": {
      "type": "integer",
      "title": "Assists"
     }
    },
    "type": "object",
    "required": [
     "player_id",
     "name",
     "minutes",
     "substitute",
     "goals",
     "assists"
    ],
    "title": "LineupPlayer"
   },
...
   "Prediction": {
    "properties": {
     "home_win": {
      "type": "number",
      "title": "Home Win"
     },
     "draw": {
      "type": "number",
      "title": "Draw"
     },
     "away_win": {
      "type": "number",
      "title": "Away Win"
     },
     "exp_home_goals": {
      "type": "number",
      "title": "Exp Home Goals"
     },
     "exp_away_goals": {
      "type": "number",
      "title": "Exp Away Goals"
     }
    },
    "type": "object",
    "required": [
     "home_win",
     "draw",
     "away_win",
     "exp_home_goals",
     "exp_away_goals"
    ],
    "title": "Prediction"
   },
...
   "TeamRef": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "logo": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Logo"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name"
    ],
    "title": "TeamRef"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  }
 }
}